fix: amount type in transaction model

+++ lib/src/shared/models/transaction.dart
@@ -4,7 +4,7 @@ class Transaction {
   int sourceAmount;
   String sourceAsset;
   String targetAsset;
-  double amount;
+  int amount;
   String type;
 
   Transaction({
@@ -24,7 +24,7 @@ class Transaction {
       sourceAmount: json["sourceAmount"] as int,
       sourceAsset: json["sourceAsset"] as String,
       targetAsset: json["targetAsset"] as String,
-      amount: json["amount"] as double,
+      amount: json["amount"] as int,
       type: json["type"] as String,
     );
   }
